feat(backend+frontend): Add endpoint to fetch new collection data and update frondent
This commit adds a new endpoint "/newcollection" in the backend to fetch
new collection data. It retrieves the products from the database,
selects the latest 8 products as the new collection, and sends them as a
response. Additonally, in the frontend, a useEffect hook is used in the
NewCollections component to fetch the new collection data upon compnent
mount and update the state accordingly.