feat(backend/frontend): Add endpoint to fetch popular proudcts in women section and update frontend
This commit introduces a new endpoint "/popularinwomen" in the backend
to retrieve popular products in the women's section. It fetches products
with the category "women" from the database, selects the top 4 popular
products, and sends them as a response. Additionaly, in the frontend,
the Popular component is updated to use a useEffect hook to fetch the
popular products upon component mount and update the state accordingly.